Use shared usdBudget extractor across proposal hook and member-vote tally
Member Vote Results was still showing $0 for proposals that don't write
their budget into proposal.budget[] in a recognizable form. The proposal
hook (useProposalJSON) handles this with a 3-step pipeline that the
server-side member-vote tally was missing.

Extracted that pipeline into a single shared helper at
ui/lib/proposals/extractUsdBudget.ts that both the client hook and the
server-side computeMemberVoteOutcome now call. Resolution order:

  1. proposal.totalBudgetUSDC (precomputed aggregate from the editor).
  2. proposal.budget[] - sum entries whose token is USD-like, accepting
     both the legacy symbol form (USD/USDC/USDT/DAI, case-insensitive)
     and the contract-address form SafeTokenForm writes today (when an
     optional stablecoinAddresses set is supplied).
  3. parseUsdBudgetFromBody - pull a USD total out of the markdown body
     (Grand Total / USD totals rows, Estimated|Total Budget callouts,
     freeform "Item - $1,234" lines).

useProposalJSON now delegates to the shared extractor (no behavioral
change for the dashboard). computeMemberVoteOutcome passes its
chain-specific stablecoin address set so the budget-cap and the
displayed numbers reflect every recognizable form.

Made-with: Cursor
